Made dlhandle_unix handle new libopenbabels
Removed the hardcoded limit to the so-version 4 of OpenBabel to enable running with development version of OB. May break things in the future. However, the current version only hardcodes it here, and doesn't check for it at compilation time leading to a very difficult to debug situation. Maybe could use some kind of CMake-variable to enable dev versions and if not, require a specific version.